requirements.txt
langchain
langchain-openai
langgraph
streamlit
aiosqlite
qdrant-client
python-dotenv
pandas
plotly
uuid6